+++ src/routes/posts/index.svelte
@@ -20,7 +20,7 @@
 	import SearchBar from '@ignatiusmb/elements/svelte/SearchBar.svelte';
 	import Pagination from '@ignatiusmb/elements/svelte/Pagination.svelte';
 	import MetaHead from '../../pages/MetaHead.svelte';
-	import PageHeader from '../../pages/PageHeader.svelte';
+	import GridView from '../../pages/GridView.svelte';
 	import Centered from '../../pages/Centered.svelte';
 	import PostCard from '../../components/PostCard.svelte';
 
@@ -37,34 +37,33 @@
 
 <MetaHead canonical="posts" title="Posts" description="Get the latest most recent posts here." />
 
-<PageHeader>
-	<h1>Posts by Mauss</h1>
+<GridView itemSize={21}>
+	<header slot="header">
+		<h1>Posts by Mauss</h1>
 
-	<SearchBar bind:query bind:filters {unique}>
-		<section>
-			<h3>Sort by</h3>
-			<label>
-				<input type="radio" bind:group={filters.sort} value="published" />
-				<span>Date published</span>
-			</label>
-			<label>
-				<input type="radio" bind:group={filters.sort} value="updated" />
-				<span>Last updated</span>
-			</label>
-		</section>
-	</SearchBar>
+		<SearchBar bind:query bind:filters {unique}>
+			<section>
+				<h3>Sort by</h3>
+				<label>
+					<input type="radio" bind:group={filters.sort} value="published" />
+					<span>Date published</span>
+				</label>
+				<label>
+					<input type="radio" bind:group={filters.sort} value="updated" />
+					<span>Last updated</span>
+				</label>
+			</section>
+		</SearchBar>
 
-	<Pagination {store} {total} {bound} />
-</PageHeader>
+		<Pagination {store} {total} {bound} />
+	</header>
 
-<main>
 	{#each sieved.slice($store * bound, $store * bound + bound) as post (post.slug)}
 		<div animate:flip={{ duration }} transition:scale|local={{ duration }}>
 			<PostCard {post} />
 		</div>
 	{/each}
-</main>
-
+</GridView>
 {#if $mobile}
 	<Centered>
 		<Pagination {store} {total} {bound} />
@@ -72,20 +71,10 @@
 {/if}
 
 <style>
-	main {
-		max-width: 84em;
-		width: 100%;
-		display: grid;
-		gap: 1em;
-		grid-template-columns: repeat(auto-fill, minmax(20em, 1fr));
-		justify-content: center;
-		padding: 0 1em;
-		margin: 0 auto;
+	h1 {
+		text-align: center;
 	}
-	main div {
+	div {
 		display: grid;
-	}
-	main :global(img:not([src])) {
-		display: none;
 	}
 </style>
